refactor: 폴더 구조 변경
_modal에는 모달의 기능에 관련된 컴포넌트를, 그 외 UI관련 부분은 _modules로 이동합니다.
