+++ src/app/(afterlogin)/presentation/upload/[id]/_component/InputSection.tsx
@@ -11,7 +11,7 @@ import Button from '@/app/_components/_elements/Button';
 import UploadPpt from './UploadPpt';
 import ControlButtons from './ControlButtons';
 import { useToastStore } from '@/store/modal';
-import SaveToast from '@/app/_components/_modules/_modal/SaveToast';
+import SaveToast from '@/app/_components/_modules/SaveToast';
 
 interface InputSectionProps {
   presentationData: PagesDataType;
