+++ src/app/_components/_modules/_modal/Modal.tsx
@@ -3,8 +3,8 @@
 import { useModalStore } from '@/store/modal';
 
 import styles from './Modal.module.scss';
-import AlertModal from './AlertModal';
-import ConfirmModal from './ConfirmModal';
+import AlertModal from '../AlertModal';
+import ConfirmModal from '../ConfirmModal';
 
 const Modal = () => {
   const { isOpen, modalData } = useModalStore();
